requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
slowapi>=0.1.9
pydantic>=2.7.0
